+++ R/Boot_scans.R
@@ -7,6 +7,13 @@
 #' @param method Character scalar, specify if the function should use a whole group or a focal scan sampling method (or both).
 #' @param focal.list Character vector, indicate the list of focals to consider throughout the scans.
 #' @param scaled logical, specifies if adjacency data should be scaled by sampling effort.
+#' @param obs.prob either :
+#' \itemize{
+#'  \item{"a dyad observation obs.probability matrix"}{of same dimension as Adj}
+#'  \item{"a dyad observation vector"}{subsetted similarly as Adj (through the non.diagonal() function for instance)}
+#'  \item{"a general dyad observation obs.probability"}{should be in [0,1], assumed to be the case when only one value is inputed)}
+#' }
+#' @param keep logical. Relevant if group scans are performed. Indicate if the original "theoretical" group scan should be kept track of.
 #' @param mode Character scalar, specifies how igraph should interpret the supplied matrix. See also the weighted argument, the interpretation depends on that too. Possible values are: directed, undirected, upper, lower, max, min, plus. See details \link[igraph]{graph_from_adjacency_matrix}.
 #' @param output Character scalar, specify if the function should return the list of scans, or reduce them into the bootstrapped adjacency matrix
 #' @param n.cores number of threads to use while performingh the bootstrap
@@ -41,13 +48,14 @@
 #' #              method = "both",mode = "directed",output = "adj",n.cores = 1)
 #' # )
 #' # system.time( #multi threaded (with number of thread existing - 1)
-#' #   Boot_scans(Adj,10,total_scan = 42,focal.list = focal.list,scaled = TRUE,
+#' #   Boot_scans(Adj,10,total_scan = 42,focal.list = focal.list,
+#' #             scaled = TRUE,obs.prob=0.7,keep=TRUE,
 #' #             method = "both",mode = "directed",output = "adj")
 #' # )
 Boot_scans<- function(Adj,n.boot,total_scan,method=c("group","focal","both"),
                       focal.list=NULL,scaled=FALSE,obs.prob=1,keep=FALSE,
                       mode = c("directed", "undirected", "max","min", "upper", "lower", "plus"),
-                      output=c("list","adjacency"),n.cores=(parallel::detectCores()-1),cl=NULL){
+                      output=c("list","adjacency","all"),n.cores=(parallel::detectCores()-1),cl=NULL){
   b<-NULL; #irrelevant bit of code, only to remove annoying note in R CMD Check...
   method<- match.arg(method)
   output<- match.arg(output)
@@ -64,7 +72,7 @@ Boot_scans<- function(Adj,n.boot,total_scan,method=c("group","focal","both"),
     doSNOW::registerDoSNOW(cl);on.exit(snow::stopCluster(cl))
   }
 
-  pbapply::pblapply(
+  Bootstrap<- pbapply::pblapply(
     1:n.boot,
     function(b){
       iterate_scans(Adj = Adj,total_scan = total_scan,
@@ -72,4 +80,5 @@ Boot_scans<- function(Adj,n.boot,total_scan,method=c("group","focal","both"),
                     method = method,mode = mode,output = output,n.cores = n.cores,cl=cl)
     }
   )
+  Bootstrap_add.attributes(Bootstrap = Bootstrap,method = method,keep = keep,mode = mode,output = output)
 }